Validations trait, and exception handler modified

+++ app/Services/ExchangeRateService.php
@@ -11,11 +11,11 @@ class ExchangeRateService implements CalculatorServiceInterface
     private const USD_EXCHANGE_RATE = 1.0;
 
     /**
-     * @param int $amount
+     * @param float $amount
      * @param Currency $currency
      * @return float
      */
-    public function calculate(int $amount, Currency $currency): float
+    public function calculate(float $amount, Currency $currency): float
     {
         return round(self::USD_EXCHANGE_RATE * $amount/$currency->exchangeRate->rate, 2);
     }
